requirements.txt
# Core dependencies
numpy>=1.19.5
pandas>=1.3.0
requests>=2.25.1

# Video analysis
opencv-python>=4.5.3
deepface>=0.0.75
moviepy>=1.0.3
openai-whisper>=20231117
ffmpeg-python>=0.2.0

# Data visualization
matplotlib>=3.4.2
seaborn>=0.11.1

# API clients
google-api-python-client>=2.0.0
polygon-api-client>=1.0.0

# Web scraping
beautifulsoup4>=4.9.3
trafilatura>=0.8.2

# Sentiment analysis
vaderSentiment>=3.3.2
transformers>=4.5.0
torch>=1.9.0
